I fixed your files
I noticed you haven't update your files in a while and I really like this app and would like to see this app thrive.  So I took the liberty of helping you out and updating it to work with the latest version of ios.  I noticed you haven't update these files in 3 years and would like to know if I could take over and contine updating it and publish to the app store so people can keep using it and this can still be an amazing app.